+++ src/parser/node/module.rs
@@ -13,7 +13,7 @@
 //////////////
 
 use super::GenericNode;
-use runtime::{Runtime, NodeOutput, Job};
+use runtime::{Runtime, Execute, Job};
 use parser::node;
 
 /////////////
@@ -59,8 +59,8 @@ impl Module {
     }
 }
 
-impl NodeOutput for Module {
-    fn output(&self, runtime: &Runtime, job: &mut Job) {
+impl Execute for Module {
+    fn execute(&self, runtime: &Runtime, job: &mut Job) {
         for node in &self.nodes {
              node.run(runtime, job)
         }