+++ src/parser/node/print.rs
@@ -3,7 +3,7 @@
 // For the copyright and license information, please view the LICENSE
 // file that was distributed with this source code.
 
-/// Represents a node that outputs an expression.
+/// Represents a node that executes an expression.
 ///
 /// @author Colin Kiegel <kiegel@gmx.de>
 
@@ -13,7 +13,7 @@
 //////////////
 
 use super::GenericNode;
-use runtime::{Runtime, NodeOutput, Job};
+use runtime::{Runtime, Execute, Job};
 use lexer::token::stream::Position;
 use parser::api::Node;
 use std::clone::Clone;
@@ -37,8 +37,8 @@ impl Print {
     }
 }
 
-impl NodeOutput for Print {
-    fn output(&self, runtime: &Runtime, job: &mut Job) {
+impl Execute for Print {
+    fn execute(&self, runtime: &Runtime, job: &mut Job) {
         for node in &self.nodes {
              node.run(runtime, job)
         }